+++ sif/sif/apps/inventario/views.py
@@ -1,0 +1,1 @@
+# Create your views here.
